# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DentOzone clinic operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{patientId}/visits:
    parameters:
      - in: path
        name: patientId
        required: true
        schema: { type: integer }
    get:
      operationId: listPatientVisits
      tags: [clinical]
      summary: List a patient's visits
      responses:
        "200":
          description: Patient visits
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Visit" }
    post:
      operationId: createPatientVisit
      tags: [clinical]
      summary: Record a patient visit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VisitInput" }
      responses:
        "201":
          description: Recorded visit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visit" }
  /visits/{visitId}:
    parameters:
      - in: path
        name: visitId
        required: true
        schema: { type: integer }
    get:
      operationId: getVisit
      tags: [clinical]
      summary: Get a visit
      responses:
        "200":
          description: Visit details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visit" }
    patch:
      operationId: updateVisit
      tags: [clinical]
      summary: Update a visit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VisitUpdate" }
      responses:
        "200":
          description: Updated visit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Visit" }
  /visits/{visitId}/diagnoses:
    parameters:
      - in: path
        name: visitId
        required: true
        schema: { type: integer }
    post:
      operationId: addVisitDiagnosis
      tags: [clinical]
      summary: Record a diagnosis for a visit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DiagnosisInput" }
      responses:
        "201":
          description: Recorded diagnosis
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Diagnosis" }
  /visits/{visitId}/procedures:
    parameters:
      - in: path
        name: visitId
        required: true
        schema: { type: integer }
    post:
      operationId: addVisitProcedure
      tags: [clinical]
      summary: Record a procedure for a visit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProcedureInput" }
      responses:
        "201":
          description: Recorded procedure
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Procedure" }
...
components:
  schemas:
...
    Visit:
      type: object
      required: [id, patientId, patientName, appointmentId, doctorStaffId, doctorName, status, complaint, notes, occurredAt, createdAt, updatedAt, isDemo, diagnoses, procedures]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: string }
        appointmentId: { type: ["integer", "null"] }
        doctorStaffId: { type: integer }
        doctorName: { type: string }
        status: { type: string, enum: [draft, completed] }
        complaint: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        occurredAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        isDemo: { type: boolean }
        diagnoses:
          type: array
          items: { $ref: "#/components/schemas/Diagnosis" }
        procedures:
          type: array
          items: { $ref: "#/components/schemas/Procedure" }
    VisitInput:
      type: object
      properties:
        appointmentId: { type: integer }
        complaint: { type: string }
        notes: { type: string }
    VisitUpdate:
      type: object
      properties:
        complaint: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        status: { type: string, enum: [draft, completed] }
    Diagnosis:
      type: object
      required: [id, visitId, label, code, toothCode, notes, recordedAt, recordedByName]
      properties:
        id: { type: integer }
        visitId: { type: integer }
        label: { type: string }
        code: { type: ["string", "null"] }
        toothCode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        recordedAt: { type: string, format: date-time }
        recordedByName: { type: string }
    DiagnosisInput:
      type: object
      required: [label]
      properties:
        label: { type: string }
        code: { type: ["string", "null"] }
        toothCode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    Procedure:
      type: object
      required: [id, visitId, label, toothCode, notes, recordedAt, recordedByName]
      properties:
        id: { type: integer }
        visitId: { type: integer }
        label: { type: string }
        toothCode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        recordedAt: { type: string, format: date-time }
        recordedByName: { type: string }
    ProcedureInput:
      type: object
      required: [label]
      properties:
        label: { type: string }
        toothCode: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
...
  securitySchemes:
    staffSession:
      type: http
      scheme: bearer
      bearerFormat: DentOzone staff session
...